+++ desktop/src/main/java/bisq/desktop/primary/main/content/user/user_profile/create/step1/GenerateNewProfileStep1Model.java
@@ -17,7 +17,7 @@
 
 package bisq.desktop.primary.main.content.user.user_profile.create.step1;
 
-import bisq.desktop.primary.overlay.onboarding.profile.GenerateProfileModel;
+import bisq.desktop.primary.overlay.onboarding.create_profile.GenerateProfileModel;
 import lombok.Getter;
 
 @Getter
