+++ desktop/src/main/java/bisq/desktop/primary/main/content/user/user_profile/create/step2/GenerateNewProfileStep2Controller.java
@@ -22,7 +22,7 @@ import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.common.view.InitWithDataController;
 import bisq.desktop.components.robohash.RoboHash;
 import bisq.desktop.primary.overlay.OverlayController;
-import bisq.desktop.primary.overlay.onboarding.profile.KeyPairAndId;
+import bisq.desktop.primary.overlay.onboarding.create_profile.KeyPairAndId;
 import bisq.identity.Identity;
 import bisq.security.pow.ProofOfWork;
 import bisq.user.identity.UserIdentityService;
