+++ desktop/src/main/java/bisq/desktop/primary/main/content/user/user_profile/create/step2/GenerateNewProfileStep2Model.java
@@ -18,7 +18,7 @@
 package bisq.desktop.primary.main.content.user.user_profile.create.step2;
 
 import bisq.desktop.common.view.Model;
-import bisq.desktop.primary.overlay.onboarding.profile.KeyPairAndId;
+import bisq.desktop.primary.overlay.onboarding.create_profile.KeyPairAndId;
 import bisq.identity.Identity;
 import bisq.security.pow.ProofOfWork;
 import javafx.beans.property.*;
