+++ desktop/src/main/java/bisq/desktop/primary/overlay/onboarding/OnboardingController.java
@@ -24,8 +24,8 @@ import bisq.desktop.common.view.NavigationController;
 import bisq.desktop.common.view.NavigationTarget;
 import bisq.desktop.primary.overlay.OverlayController;
 import bisq.desktop.primary.overlay.onboarding.bisq2.Bisq2IntroController;
+import bisq.desktop.primary.overlay.onboarding.create_profile.GenerateProfileController;
 import bisq.desktop.primary.overlay.onboarding.password.OnboardingPasswordController;
-import bisq.desktop.primary.overlay.onboarding.profile.GenerateProfileController;
 import javafx.application.Platform;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
