+++ desktop/src/main/java/bisq/desktop/primary/overlay/onboarding/create_profile/GenerateProfileController.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.primary.overlay.onboarding.profile;
+package bisq.desktop.primary.overlay.onboarding.create_profile;
 
 import bisq.application.DefaultApplicationService;
 import bisq.desktop.common.threading.UIThread;
